Savegame/mapgen signs

+++ src/ui.cpp
@@ -75,7 +75,7 @@ uimenu::uimenu(bool cancelable, const char *mes,
         shift_retval = 1;
         return_invalid = cancelable;
 
-        for (int i = 0; i < options.size(); i++) {
+        for (size_t i = 0; i < options.size(); i++) {
             entries.push_back(uimenu_entry(i, true, MENU_AUTOASSIGN, options[i] ));
         }
         query();
